+++ tests/test.lunaris.storeAndHook.js
@@ -4610,5 +4610,25 @@ describe('Lunaris hooks', () => {
       should(lunaris._stores['store1'].hooks.get[0]).eql(_handler2);
       delete lunaris._stores['store1'];
     });
+
+    it('should remove one handler from a list of handlers', () => {
+      var _handler1             = function handler1 () {};
+      var _handler2             = function handler1 () {};
+      lunaris._stores['store1'] = { hooks : [] };
+      lunaris.hook('get@store1', _handler1);
+      lunaris.hook('get@store1', _handler2);
+      should(lunaris._stores['store1'].hooks).be.an.Array();
+      should(lunaris._stores['store1'].hooks.get).have.length(2);
+      should(lunaris._stores['store1'].hooks.get[0]).be.a.Function();
+      should(lunaris._stores['store1'].hooks.get[0]).eql(_handler1);
+      should(lunaris._stores['store1'].hooks.get[1]).be.a.Function();
+      should(lunaris._stores['store1'].hooks.get[1]).eql(_handler2);
+      lunaris.removeHook('get@store1', _handler1);
+      should(lunaris._stores['store1'].hooks).be.an.Array();
+      should(lunaris._stores['store1'].hooks.get).have.length(1);
+      should(lunaris._stores['store1'].hooks.get[0]).be.a.Function();
+      should(lunaris._stores['store1'].hooks.get[0]).eql(_handler2);
+      delete lunaris._stores['store1'];
+    });
   });
 });